move update grade & get transcript endpoints to StudentController
