+++ src/student_management_api/Controllers/CourseEnrollmentController.cs
@@ -91,55 +91,4 @@ public class CourseEnrollmentController : ControllerBase
             return BadRequest(new ErrorResponse<string>(status: 400, message: _localizer["invalid_action_specified_Use_register_or_unregister"]));
         }
     }
-
-    [HttpPut("update-grade")]
-    public async Task<IActionResult> UpdateStudentGrade([FromBody] UpdateStudentGradeRequest request)
-    {
-        if (!ModelState.IsValid)
-        {
-            return BadRequest(new ErrorResponse<ModelStateDictionary>(status: 400, message: _localizer["invalid_input"], details: ModelState));
-        }
-
-        using (_logger.BeginScope("UpdateStudentGrade request"))
-        {
-            _logger.LogInformation("Updating student grade with request: {@Request}", request);
-
-            try
-            {
-                await _courseEnrollmentService.UpdateStudentGrade(request.StudentId!, request.CourseId!, request.Grade);
-                _logger.LogInformation("Successfully updated student grade");
-
-                return Ok(new { message = _localizer["successfully_updated_student_grade"].Value });
-            }
-            catch
-            {
-                _logger.LogWarning("Failed to update student grade");
-                throw;
-            }
-        }
-    }
-
-    [HttpGet("transcript/{student_id}")]
-    public async Task<IActionResult> GetStudentTranscriptById(string student_id)
-    {
-        using (_logger.BeginScope("GetStudentTranscriptById request for StudentId: {StudentId}", student_id))
-        {
-            // Load HTML template
-            var culture = CultureInfo.CurrentUICulture.TwoLetterISOLanguageName;
-            var templatePath = Path.Combine(Directory.GetCurrentDirectory(), "Templates", $"transcript_template-{culture}.html");
-            var htmlTemplate = await System.IO.File.ReadAllTextAsync(templatePath);
-
-            _logger.LogInformation("Fetching transcript for student with ID: {StudentId}", student_id);
-            var transcriptStream = await _courseEnrollmentService.GetTranscriptOfStudentById(student_id, htmlTemplate);
-            if (transcriptStream == null)
-            {
-                _logger.LogWarning("Transcript not found for student with ID: {StudentId}", student_id);
-                return NotFound(new ErrorResponse<string>(status: 404, message: $"{_localizer["transcript_not_found"]}, ID: {student_id}"));
-            }
-
-            _logger.LogInformation("Transcript fetched successfully for student with ID: {StudentId}", student_id);
-
-            return File(transcriptStream, "application/pdf", $"{student_id}_transcript.pdf");
-        }
-    }
 }
